+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - Added header (navigation bar) and footer (credits) to every page
 - Released Profile page
 - Released 400m minigame
-- Released API
+- Released API routes
+- Applied limiter to limit query load and prevent bruteforce attacks
 
 ### Fixes
 - Fixed register / login / logout buttons, forms, submits